+++ src/main/java/com/grepp/team07/app/model/product/ProductRepository.java
@@ -14,35 +14,6 @@ import org.apache.ibatis.annotations.Update;
 @Mapper
 public interface ProductRepository {
 
-    List<ProductDto> selectAll();
-
-    ProductDto selectById(Integer id);
-
-    List<ProductDto> searchByKeyword(String keyword);
-
-    @Insert("insert into product(name, price, info, brand)"
-        + "values (#{name}, #{price}, #{info}, #{brand})")
-    @Options(useGeneratedKeys = true, keyProperty = "productId")
-    void insert(ProductDto productDto);
-
-    @Select("insert into product_img (product_id, original_file_name, rename_file_name, save_path)"
-        + "values (#{productId}, #{originalFileName}, #{renameFileName}, #{savePath})")
-    void insertImage(ProductImgDto productImgDto);
-
-    @Update("update product set name=#{name}, price=#{price}, info=#{info}, brand=#{brand} where product_id=#{productId}")
-    boolean update(ProductDto productDto);
-
-    @Update("UPDATE product SET count = count + 1 WHERE product_id = #{productId}")
-    void increment(@Param("productId") Integer productId);
-
-    @Update("UPDATE product SET count = count - 1 WHERE product_id = #{productId} AND count > 0")
-    void decrement(@Param("productId") Integer productId);
-
-    @Update("UPDATE product SET deleted_at = NOW() WHERE product_id = #{productId}")
-    void delete(@Param("productId") Integer productId);
-
-    @Update("UPDATE product_img SET activated = 0 WHERE product_id = #{productId}")
-    void deleteImage(@Param("productId") Integer productId);
     List<ProductDto> findByKeyword(String keyword);
 
     List<ProductDto> bestProduct();
