+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-upload.path=C:/backend/upload/
+upload.path=/Users/eonwy/Programmers/backend/f_spring/upload
 spring.datasource.url=jdbc:mysql://localhost:3306/team07?useUnicode=true&characterEncoding=utf8
 spring.datasource.username=bm
 spring.datasource.password=123qwe!@#
